fixed warning

+++ headers/tester.h
@@ -1,19 +1,14 @@
 #ifndef TESTER_H
 #define TESTER_H
 
-#include <stdlib.h>
-#include <time.h>
-#include <string.h>
-
 #include <solver.h>
-#include <colors.h>
 
 /*! \brief Runs n_tests from array tests. Returns number of passed tests.
 \param tests array of tests
 \param n_tests number of tests
 \return number of passed tests
  */
-int testSolver(struct Equation* tests, int n_tests);
+int testSolver(struct Equation* tests, size_t n_tests);
 
 /*! \brief Runs n_tests random tests. Returns number of passed tests.
 \param n_tests number of tests
@@ -23,11 +18,11 @@ int testRand(int n_tests);
 
 long fileSize(FILE *file);
 
-int nLines(const char* s);
+size_t nLines(const char* s);
 
-struct Equation* getTests(const char* s, int n_lines, int* error);
+struct Equation* getTests(const char* s, size_t n_lines, int* error);
 
-char* readFile(FILE* file, int *n_chars, int *error);
+char* readFile(FILE* file, size_t *n_chars, int *error);
 
 /*! \brief (old docs) Reads tests from specified file in an array.
 
@@ -37,7 +32,7 @@ File should contain n_tests lines. Each line should contain 6 mumbers and repres
 * \param n_tests number of tests
 * \return 0 if errors occur nored, nonzero otherwise
  */
-struct Equation* readTests(const char* file_name, int *n_tests, int *error);
+struct Equation* readTests(const char* file_name, size_t *n_tests, int *error);
 
 /*! \brief Generates a random double from -rand_max to rand_max
 \param rand_max boundary for a generated number
@@ -49,6 +44,6 @@ double genRand(int rand_max);
 \param n_tests int that will store number of tests
 \return number of passed tests
  */
-int readAndTest(const char* file, int *n_tests);
+int readAndTest(const char* file, size_t *n_tests);
 
 #endif // TESTER_H
